posix: use new uid/gid getters in Process for getuid/setuid et al

+++ src/subsys/posix/system-syscalls.h
@@ -95,7 +95,7 @@ int posix_arch_prctl(int code, unsigned long addr);
 int posix_pause();
 
 int posix_setgroups(size_t size, const gid_t *list);
-int posix_getgroups(int size, gid_t *list);
+int posix_getgroups(size_t size, gid_t *list);
 
 int posix_getrlimit(int resource, struct rlimit *rlim);
 int posix_setrlimit(int resource, const struct rlimit *rlim);
